refactor: move sharingGroup.js one level up
