+++ csaf_2_1/mandatoryTests/mandatoryTest_6_1_39.js
@@ -1,5 +1,5 @@
 import Ajv from 'ajv/dist/jtd.js'
-import { MAX_UUID, NIL_UUID } from './sharingGroup.js'
+import { MAX_UUID, NIL_UUID } from '../sharingGroup.js'
 
 const ajv = new Ajv()
