+++ csaf_2_1/sharingGroup.js
